Link to file with pre-trained weights to README
Hi, thanks for releasing the code and weights! Given that there's two issues asking about weights (#6 and #10) and I also went to the issues for asking after not seeing them mentioned anywhere in the README, I think this is confusing and worth linking from the README. So where we go.